feat: implement support for single custom prefix
This allows for an administrator to set a prefix that will
be used for all authenticators configured.

It can be useful when there's a need to use an alternate login
method while keeping the same username.

**WARNING: This can have security implications !!!**

If two different users, using each a different login service
have the same username returned (e.g. User1 on GitLab gets UserXYZ
and User2 on GitHub also get UserXYZ, then they will have access
to the same account on the JupyterHub deployment.